[GH-3] refactor: Added nest-like interface structure
