+++ server/package.json
@@ -15,7 +15,6 @@
         "body-parser": "^1.19.0",
         "config": "^3.0.1",
         "cors": "^2.8.5",
-        "express": "^4.16.3",
         "lodash": "^4.17.15",
         "reflect-metadata": "^0.1.13",
         "rxjs": "^6.5.2",
