+++ server/src/artists/interfaces/artist.interface.ts
@@ -1,0 +1,7 @@
+import ITrack from './track.interface';
+
+export default interface IArtist {
+  name: string;
+  wikipediaDescription: string;
+  topTracks?: ITrack[];
+}
