+++ server/src/artists/interfaces/cache.interface.ts
@@ -1,0 +1,4 @@
+export interface ICache {
+  value: string;
+  timestamp?: number;
+}
