+++ server/src/artists/interfaces/spotify.interface.ts
@@ -1,15 +1,3 @@
-export interface ITrack {
-  title: string;
-  spotifyLink?: string;
-  albumCover?: string;
-}
-
-export interface IArtist {
-  name: string;
-  wikipediaDescription: string;
-  topTracks?: ITrack[];
-}
-
 export interface ISpotify {
   clientCredentialsGrant: () => Promise<{ body: object }>;
   setAccessToken: (token: string) => void;
@@ -20,9 +8,4 @@ export interface ISpotify {
 export interface ISpotifyConfig {
   numberOfBestSongs: number;
   locale: string;
-}
-
-export interface ICache {
-  value: string;
-  timestamp?: number;
 }
