+++ server/src/artists/interfaces/track.interface.ts
@@ -1,0 +1,5 @@
+export default interface ITrack {
+  title: string;
+  spotifyLink?: string;
+  albumCover?: string;
+}
